+++ src/services/UserService/GymInnowise.UserService.API/Controllers/PersonalGoalController.cs
@@ -1,6 +1,8 @@
-﻿using GymInnowise.UserService.Logic.Interfaces;
+﻿using GymInnowise.UserService.API.Authorization;
+using GymInnowise.UserService.Logic.Interfaces;
 using GymInnowise.UserService.Shared.Dtos.RequestModels.Creates;
 using GymInnowise.UserService.Shared.Dtos.RequestModels.Updates;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GymInnowise.UserService.API.Controllers
@@ -10,12 +12,16 @@ namespace GymInnowise.UserService.API.Controllers
     public class PersonalGoalController : ControllerBase
     {
         private readonly IPersonalGoalService _personalGoalService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public PersonalGoalController(IPersonalGoalService personalGoalService)
+        public PersonalGoalController(IPersonalGoalService personalGoalService,
+            IAuthorizationService authorizationService)
         {
             _personalGoalService = personalGoalService;
+            _authorizationService = authorizationService;
         }
 
+        [Authorize]
         [HttpGet("{ownerId}")]
         public async Task<IActionResult> GetPersonalGoalsAsync(Guid ownerId)
         {
@@ -24,9 +30,18 @@ namespace GymInnowise.UserService.API.Controllers
             return Ok(goals);
         }
 
+        [Authorize]
         [HttpPut]
         public async Task<IActionResult> UpdatePersonalGoalAsync([FromBody] UpdatePersonalGoalRequest request)
         {
+            var owner = await _personalGoalService.GetOwnerAsync(request.Id);
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(User, owner, PolicyNames.OwnerOrAdmin);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             var updateResult = await _personalGoalService.UpdatePersonalGoalAsync(request);
 
             return updateResult.Match<IActionResult>(
@@ -34,10 +49,21 @@ namespace GymInnowise.UserService.API.Controllers
                 _ => NotFound());
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> CreatePersonalGoalAsync([FromBody] CreatePersonalGoalRequest request)
         {
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(User, request.Owner, PolicyNames.OwnerOrAdmin);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             await _personalGoalService.CreatePersonalGoalAsync(request);
+            var userAccountId = User.Claims.FirstOrDefault(c => c.Type == "accountId")?.Value;
+
+            Console.WriteLine($"Account id is {userAccountId} and should be {request.Owner}");
 
             return Created();
         }
@@ -45,6 +71,14 @@ namespace GymInnowise.UserService.API.Controllers
         [HttpDelete("goalId")]
         public async Task<IActionResult> RemovePersonalGoalAsync(Guid goalId)
         {
+            var owner = await _personalGoalService.GetOwnerAsync(goalId);
+            var authorizationResult =
+                await _authorizationService.AuthorizeAsync(User, owner, PolicyNames.OwnerOrAdmin);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             await _personalGoalService.RemovePersonalGoalAsync(goalId);
 
             return NoContent();
